requirements.txt
# AI Teaching Assistant - Dependencies
# Install with: pip install -r requirements.txt

# Azure OpenAI SDK
openai>=1.0.0

# Environment variable management
python-dotenv>=1.0.0

# Web UI framework (used in step 9)
streamlit>=1.28.0

# Utility libraries
requests>=2.31.0

